all: make scroll axis handling more robust

+++ src/cli/seat_test.rs
@@ -190,14 +190,14 @@ async fn run(seat_test: Rc<SeatTest>) {
     let st = seat_test.clone();
     AxisFrame::handle(tc, se, ps.clone(), move |ps, ev| {
         let source = ps.source.take();
-        let px_x = ps.px[HORIZONTAL_SCROLL as usize].take();
-        let px_y = ps.px[VERTICAL_SCROLL as usize].take();
-        let stop_x = ps.stop[HORIZONTAL_SCROLL as usize].take();
-        let stop_y = ps.stop[VERTICAL_SCROLL as usize].take();
-        let v120_x = ps.v120[HORIZONTAL_SCROLL as usize].take();
-        let v120_y = ps.v120[VERTICAL_SCROLL as usize].take();
-        let inverted_x = ps.inverted[HORIZONTAL_SCROLL as usize].get();
-        let inverted_y = ps.inverted[VERTICAL_SCROLL as usize].get();
+        let px_x = ps.px[HORIZONTAL_SCROLL].take();
+        let px_y = ps.px[VERTICAL_SCROLL].take();
+        let stop_x = ps.stop[HORIZONTAL_SCROLL].take();
+        let stop_y = ps.stop[VERTICAL_SCROLL].take();
+        let v120_x = ps.v120[HORIZONTAL_SCROLL].take();
+        let v120_y = ps.v120[VERTICAL_SCROLL].take();
+        let inverted_x = ps.inverted[HORIZONTAL_SCROLL].get();
+        let inverted_y = ps.inverted[VERTICAL_SCROLL].get();
         if all || ev.seat == seat {
             if all {
                 print!("Seat: {}, ", st.name(ev.seat));
