+++ src/ifs/wl_seat/wl_pointer.rs
@@ -18,8 +18,8 @@ const ROLE: u32 = 0;
 pub(super) const RELEASED: u32 = 0;
 pub const PRESSED: u32 = 1;
 
-pub const VERTICAL_SCROLL: u32 = 0;
-pub const HORIZONTAL_SCROLL: u32 = 1;
+pub const VERTICAL_SCROLL: usize = 0;
+pub const HORIZONTAL_SCROLL: usize = 1;
 
 pub const WHEEL: u32 = 0;
 pub const FINGER: u32 = 1;
